+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/gateway
 
+## 2.0.0-beta.90
+
+### Patch Changes
+
+- cbf52cd: feat: expose raw finish reason
+- Updated dependencies [cbf52cd]
+  - @ai-sdk/provider@3.0.0-beta.30
+  - @ai-sdk/provider-utils@4.0.0-beta.57
+
 ## 2.0.0-beta.89
 
 ### Patch Changes