+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/mcp
 
+## 1.0.0-beta.44
+
+### Patch Changes
+
+- Updated dependencies [cbf52cd]
+  - @ai-sdk/provider@3.0.0-beta.30
+  - @ai-sdk/provider-utils@4.0.0-beta.57
+
 ## 1.0.0-beta.43
 
 ### Patch Changes